requirements.txt
selenium
beautifulsoup4
mysql-connector-python